+++ python/src/scipp/plotting/controller.py
@@ -657,3 +657,10 @@ class PlotController:
         Make a map from dim to axis
         """
         return {self.axes[key]: key for key in self._get_xyz_axes()}
+
+    def redraw(self):
+        """
+        Update the model data dicts and re-draw the figure.
+        """
+        self.model.update_data_arrays()
+        self.update_data()